Added player unit and bullets
Player unit and bullets have a heading as well as a position.
Player can shoot bullets with a reload speed of 2 frames
Bullets spawn slightly in front of the player and have the same heading

Bullets disappear when they either hit an enemy unit or a solid map tile
Bullets damage enemies and spawners
Enemies and spawner without hp are destroyed and removed from the game

All timed ingame events are timed using frames, this makes it easy to speedup the simulation when I want to train the AI on the game.