Starting on the hierarchical AI
Added behaviour class which contain a neural net, queue containing the history of states/actions and an array which contains the weights of the rewards compared to the universal reward formula.
The network of the behaviour can be updated by providing the rewards of the head of the history queue.