Adjusted updateNetwork function in behaviour, this makes it so that not every behaviour stores it owns history, instead the Hierarchical AI class keeps an history which consists of a FIFO queue of states. These states contain the global input, the rewards and the chosen action. When the hierarchical AI gets the update command it will look at the top 2 states in the queue, of there are 2 states and both of these do not have a bullet flying it wil pop the first one of the queue and peek at the second. Both of them will be send to each behaviour to update their neural networks accordingly.
Bullets get a spawntime at the start, this time is not used at the moment. In the future it should be use to determine which state to update when the bullet hits something. This makes sure that hitting an enemy with a bullet rewards the action which shot the bullet, not the one that was taken at the moment the bullet hit something.